backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
SQLAlchemy>=2.0.46,<2.1
psycopg[binary]==3.3.2
python-dotenv==1.0.1
litellm==1.77.3
requests==2.32.5
aiohttp==3.12.15
croniter==6.0.0
sentry-sdk==2.31.0
